Adopt '.a3s' and '.a3t' Extensions
Adopt to the upcoming official extensions `.a3s` and `.a3t` for ALAN3
solution files and transcripts, respectively (see alan-if/alan#2), but
also preserve the old `.a3sol` and `.a3log` extensions (as secondary)
for a grace period, to allow projects that are using them enough time
to migrate.

Also change the scope of the Alan Solution and Transcript syntaxes from
`.a3sol` and `.a3log` to `.a3s` and `.a3t`, respectively, and update all
package files accordingly. (closes #1)